+++ CHANGELOG.md
@@ -80,6 +80,7 @@
 - Dev: Bumped clang-format requirement to 19. (#6236)
 - Dev: Factored out AUMID to `Version`. (#6321)
 - Dev: Silenced some warnings when compiling with clang-cl. (#6331)
+- Dev: Added some commands for forcing a relayout (and related things) in channel views. (#6342)
 
 ## 2.5.3